[Performance Refactoring]
Refactored all elements to use a static field for their color, and store the RGB values of the color in their class instance. This causes the simulation to use about half of the memory it was using before. Additionally, added garbace collector calls when clearing the matrix and when switching brush types, as the GC was not runnning as often as it should have been.